Move `/faq manual` to `/fuckup manual`
1. Added some extra info to the (very similar) `/faq update`
2. made it clear the recommended path is the normal update path from the website

The entry here also has some info on viomi and might be nice for more advanded users who don't want to revuild the firmware.
But it requires people to know what they're doing, and all in all it's better to have one true path to support nand not a multitude of different ways for people to break their valetudo install.